+++ CHANGES.md
@@ -3,7 +3,10 @@
 ## ... v0.0.6
 
   * Handle case of no versions https://ocfl.io/draft/spec/#E008
+  * Handle case of bad JSON inventory https://ocfl.io/draft/spec/#E033
   * Handle case of no manifest https://ocfl.io/draft/spec/#E041
+  * Handle case of conflicting paths https://ocfl.io/draft/spec/#E095
+
 
 ## 2020-05-05 v0.0.5